# server/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API for earthquake monitor",
        "description": "This is a simple REST API application made with Express and documented with Swagger",
        "contact": {
            "email": "bostonmathis@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://earthquake-monitor-server.herokuapp.com/"
        }
    ],
...
    "paths": {
        "/auth/register": {
            "post": {
                "tags": ["Auth"],
                "summary": "Create user",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Created user object",
                    "content": {
                        "*/*": {}
                    },
                    "required": true
                },
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
        "/auth/login": {
            "post": {
                "tags": ["Auth"],
                "summary": "Logs user into the system",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "The user name for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "The password for login in clear text",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        },
                        "content": {
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied",
                        "content": {}
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "tags": ["Users"],
                "summary": "Get user in session",
                "description": "This can only be done by the logged in user.",
                "operationId": "getUserInSession",
                "parameters": [
                    {
                        "name": "body",
                        "in": "session",
                        "description": "Get user in session",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "location": {
                                            "type": "object",
                                            "properties": {
                                                "type": {
                                                    "type": "string"
                                                },
                                                "coordinates": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number",
                                                        "format": "double"
                                                    },
                                                    "description": "longitude, latitude",
                                                    "example": [-0.1337, 51.5152],
                                                    "x-codegen-array-max-length": 2,
                                                    "minItems": 2,
                                                    "maxItems": 2
                                                }
                                            }
                                        },
                                        "_id": {
                                            "type": "string"
                                        },
                                        "username": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "password": {
                                            "type": "string"
                                        },
                                        "lastVisit": {
                                            "type": "date"
                                        },
                                        "third_party_auth": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "provider_name": {
                                                        "type": "string"
                                                    },
                                                    "provider_id": {
                                                        "type": "string"
                                                    },
                                                    "provider_data": {
                                                        "type": "object"
                                                    }
                                                }
                                            }
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "public": {
                                            "type": "boolean"
                                        }
                                    },
                                    "example": {
                                        "location": {
                                            "type": "Point",
                                            "coordinates": [
                                                -151.00415781499996,
                                                64.80808789800005
                                            ]
                                        },
                                        "_id": "628d9d189df99f41ce8c55bf",
                                        "username": "AlaskaPublic",
                                        "email": "AlaskaPublic@gmail.com",
                                        "password": "$2a$10$fTdTUKSmcFAbzr3AmeHnOuQPDcoXXwV0s2IGZDKHUFYwobBboPpzC",
                                        "third_party_auth": [],
                                        "lastVisit": "2022-05-25T03:06:08.855Z",
                                        "__v": 0,
                                        "address": "Alaska",
                                        "public": true
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {}
                    },
                    "404": {
                        "description": "User not found",
                        "content": {}
                    }
                }
            }
        },
...
    },
...
}